Refactor component containers and rename some
