+++ services/tenant-ui/frontend/src/components/connections/Invitations.vue
@@ -1,0 +1,172 @@
+<template>
+  <MainCardContent
+    :title="$t('connect.invitations.invitations')"
+    :refresh-callback="loadTable"
+  >
+    <DataTable
+      v-model:expandedRows="expandedRows"
+      v-model:filters="filter"
+      :loading="loading"
+      :value="formattedInvitations"
+      :paginator="true"
+      :rows="TABLE_OPT.ROWS_DEFAULT"
+      :rows-per-page-options="TABLE_OPT.ROWS_OPTIONS"
+      :global-filter-fields="['alias']"
+      selection-mode="single"
+      data-key="connection_id"
+      sort-field="created_at"
+      :sort-order="-1"
+      filter-display="menu"
+    >
+      <template #header>
+        <div class="flex justify-content-between">
+          <div class="flex justify-content-start">
+            <CreateContact :multi="false" class="mr-3" />
+            <CreateContact :multi="true" />
+          </div>
+          <div class="flex justify-content-end">
+            <span class="p-input-icon-left">
+              <i class="pi pi-search ml-0" />
+              <InputText
+                v-model="filter.alias.value"
+                :placeholder="$t('connect.invitations.search')"
+              />
+            </span>
+          </div>
+        </div>
+      </template>
+      <template #empty>{{ $t('common.noRecordsFound') }}</template>
+      <template #loading>{{ $t('common.loading') }}</template>
+      <Column :expander="true" header-style="width: 3rem" />
+      <Column :sortable="false" :header="$t('common.actions')">
+        <template #body="{ data }">
+          <DeleteContact :connection-id="data.connection_id" />
+          <RegenerateInvitation :connection-id="data.connection_id" />
+        </template>
+      </Column>
+      <Column
+        :sortable="true"
+        field="alias"
+        :header="$t('common.alias')"
+        filter-field="alias"
+        :show-filter-match-modes="false"
+      >
+        <template #filter="{ filterModel, filterCallback }">
+          <InputText
+            v-model="filterModel.value"
+            type="text"
+            class="p-column-filter"
+            placeholder="Search By Alias"
+            @input="filterCallback()"
+          />
+        </template>
+      </Column>
+      <Column
+        :sortable="true"
+        field="invitation_mode"
+        :header="$t('connect.table.invitationMode')"
+        filter-field="invitation_mode"
+        :show-filter-match-modes="false"
+      >
+        <template #filter="{ filterModel, filterCallback }">
+          <InputText
+            v-model="filterModel.value"
+            type="text"
+            class="p-column-filter"
+            placeholder="Search By Invitation Mode"
+            @input="filterCallback()"
+          />
+        </template>
+      </Column>
+      <Column
+        :sortable="true"
+        field="created"
+        :header="$t('connect.table.createdAt')"
+        filter-field="created"
+        :show-filter-match-modes="false"
+      >
+        <template #body="{ data }">
+          {{ data.created }}
+        </template>
+        <template #filter="{ filterModel, filterCallback }">
+          <InputText
+            v-model="filterModel.value"
+            type="text"
+            class="p-column-filter"
+            placeholder="Search By Date"
+            @input="filterCallback()"
+          />
+        </template>
+      </Column>
+      <template #expansion="{ data }">
+        <RowExpandData :id="data.connection_id" :url="API_PATH.CONNECTIONS" />
+      </template>
+    </DataTable>
+  </MainCardContent>
+</template>
+
+<script setup lang="ts">
+// Vue
+import { onMounted, ref, computed } from 'vue';
+// PrimeVue
+import Column from 'primevue/column';
+import InputText from 'primevue/inputtext';
+import DataTable from 'primevue/datatable';
+import { useToast } from 'vue-toastification';
+import { FilterMatchMode } from 'primevue/api';
+// State
+import { useContactsStore } from '@/store';
+import { storeToRefs } from 'pinia';
+// Other components
+import CreateContact from '@/components/connections/createContact/CreateContact.vue';
+import DeleteContact from '@/components/connections/editContact/DeleteContact.vue';
+import MainCardContent from '../layout/mainCard/MainCardContent.vue';
+import RegenerateInvitation from '@/components/connections/createContact/RegenerateInvitation.vue';
+import RowExpandData from '@/components/common/RowExpandData.vue';
+import { TABLE_OPT, API_PATH } from '@/helpers/constants';
+import { formatDateLong } from '@/helpers';
+
+const toast = useToast();
+
+const contactsStore = useContactsStore();
+
+const { loading, filteredInvitations } = storeToRefs(useContactsStore());
+
+const formattedInvitations = computed(() =>
+  filteredInvitations.value.map((inv) => ({
+    connection_id: inv.connection_id,
+    alias: inv.alias,
+    invitation_mode: inv.invitation_mode,
+    created: formatDateLong(inv.created_at as string),
+    created_at: inv.created_at,
+  }))
+);
+const loadTable = async () => {
+  contactsStore.listContacts().catch((err) => {
+    console.error(err);
+    toast.error(`Failure: ${err}`);
+  });
+};
+
+onMounted(async () => {
+  loadTable();
+});
+
+// necessary for expanding rows, we don't do anything with this
+const expandedRows = ref([]);
+
+const filter = ref({
+  alias: {
+    value: null,
+    matchMode: FilterMatchMode.CONTAINS,
+  },
+  invitation_mode: {
+    value: null,
+    matchMode: FilterMatchMode.CONTAINS,
+  },
+  created: {
+    value: null,
+    matchMode: FilterMatchMode.CONTAINS,
+  },
+});
+</script>
