+++ services/tenant-ui/frontend/src/components/connections/acceptInvitation/AcceptInvitation.vue
