+++ services/tenant-ui/frontend/src/components/connections/acceptInvitation/AcceptInviteForm.vue
