+++ services/tenant-ui/frontend/src/components/connections/editContact/DeleteContact.vue
