+++ services/tenant-ui/frontend/src/components/connections/messageContact/MessageContactList.vue
