+++ services/tenant-ui/frontend/src/components/issuance/CredentialDefinitions.vue
@@ -1,123 +1,118 @@
 <template>
-  <h3 class="mt-0">
-    {{ $t('configuration.schemasCreds.credentialDefinitions') }}
-  </h3>
-
-  <DataTable
-    v-model:expandedRows="expandedRows"
-    v-model:filters="filter"
-    :loading="loading"
-    :value="formattedstoredCredDefs"
-    :paginator="true"
-    :rows="TABLE_OPT.ROWS_DEFAULT"
-    :rows-per-page-options="TABLE_OPT.ROWS_OPTIONS"
-    selection-mode="single"
-    data-key="cred_def_id"
-    sort-field="created_at"
-    :sort-order="-1"
-    filter-display="menu"
+  <MainCardContent
+    :title="$t('configuration.schemasCreds.credentialDefinitions')"
+    :refresh-callback="loadTable"
   >
-    <template #header>
-      <div class="flex justify-content-between">
-        <div class="flex justify-content-start"></div>
-        <div class="flex justify-content-end">
-          <span class="p-input-icon-left schema-search">
-            <i class="pi pi-search" />
-            <InputText
-              v-model="filter.global.value"
-              placeholder="Search Cred Defs"
-            />
-          </span>
-          <Button
-            icon="pi pi-refresh"
-            class="p-button-rounded p-button-outlined"
-            title="Refresh Table"
-            @click="loadTable"
-          />
+    <DataTable
+      v-model:expandedRows="expandedRows"
+      v-model:filters="filter"
+      :loading="loading"
+      :value="formattedstoredCredDefs"
+      :paginator="true"
+      :rows="TABLE_OPT.ROWS_DEFAULT"
+      :rows-per-page-options="TABLE_OPT.ROWS_OPTIONS"
+      selection-mode="single"
+      data-key="cred_def_id"
+      sort-field="created_at"
+      :sort-order="-1"
+      filter-display="menu"
+    >
+      <template #header>
+        <div class="flex justify-content-between">
+          <div class="flex justify-content-start" />
+          <div class="flex justify-content-end">
+            <span class="p-input-icon-left">
+              <i class="pi pi-search" />
+              <InputText
+                v-model="filter.global.value"
+                placeholder="Search Cred Defs"
+              />
+            </span>
+          </div>
         </div>
-      </div>
-    </template>
-    <template #empty>{{ $t('common.noRecordsFound') }}</template>
-    <template #loading>{{ $t('common.loading') }}</template>
-    <Column :expander="true" header-style="width: 3rem" />
-    <Column :sortable="false" header="Actions">
-      <template #body="{ data }">
-        <Button
-          title="Delete Credential Definition"
-          icon="pi pi-trash"
-          class="p-button-rounded p-button-icon-only p-button-text"
-          @click="deleteCredDef($event, data.cred_def_id)"
+      </template>
+      <template #empty>{{ $t('common.noRecordsFound') }}</template>
+      <template #loading>{{ $t('common.loading') }}</template>
+      <Column :expander="true" header-style="width: 3rem" />
+      <Column :sortable="false" header="Actions">
+        <template #body="{ data }">
+          <Button
+            title="Delete Credential Definition"
+            icon="pi pi-trash"
+            class="p-button-rounded p-button-icon-only p-button-text"
+            @click="deleteCredDef($event, data.cred_def_id)"
+          />
+        </template>
+      </Column>
+      <Column
+        :sortable="true"
+        field="cred_def_id"
+        header="ID"
+        filter-field="cred_def_id"
+        :show-filter-match-modes="false"
+      >
+        <template #filter="{ filterModel, filterCallback }">
+          <InputText
+            v-model="filterModel.value"
+            type="text"
+            class="p-column-filter"
+            placeholder="Search By ID"
+            @input="filterCallback()"
+          />
+        </template>
+      </Column>
+      <Column
+        :sortable="true"
+        field="schema_id"
+        header="Schema ID"
+        filter-field="schema_id"
+        :show-filter-match-modes="false"
+      >
+        <template #filter="{ filterModel, filterCallback }">
+          <InputText
+            v-model="filterModel.value"
+            type="text"
+            class="p-column-filter"
+            placeholder="Search By Schema ID"
+            @input="filterCallback()"
+          />
+        </template>
+      </Column>
+      <Column :sortable="true" field="support_revocation" header="Revokable">
+        <template #body="{ data }">
+          <span v-if="data.support_revocation">
+            <i class="pi pi-check-circle"></i>
+          </span>
+        </template>
+      </Column>
+      <Column
+        :sortable="true"
+        field="created"
+        header="Created at"
+        filter-field="created"
+        :show-filter-match-modes="false"
+      >
+        <template #body="{ data }">
+          {{ data.created }}
+        </template>
+        <template #filter="{ filterModel, filterCallback }">
+          <InputText
+            v-model="filterModel.value"
+            type="text"
+            class="p-column-filter"
+            placeholder="Search By Time"
+            @input="filterCallback()"
+          />
+        </template>
+      </Column>
+      <template #expansion="{ data }">
+        <RowExpandData
+          :id="data.cred_def_id"
+          :url="API_PATH.CREDENTIAL_DEFINITION_STORAGE"
         />
       </template>
-    </Column>
-    <Column
-      :sortable="true"
-      field="cred_def_id"
-      header="ID"
-      filter-field="cred_def_id"
-      :show-filter-match-modes="false"
-    >
-      <template #filter="{ filterModel, filterCallback }">
-        <InputText
-          v-model="filterModel.value"
-          type="text"
-          class="p-column-filter"
-          placeholder="Search By ID"
-          @input="filterCallback()"
-        />
-      </template>
-    </Column>
-    <Column
-      :sortable="true"
-      field="schema_id"
-      header="Schema ID"
-      filter-field="schema_id"
-      :show-filter-match-modes="false"
-    >
-      <template #filter="{ filterModel, filterCallback }">
-        <InputText
-          v-model="filterModel.value"
-          type="text"
-          class="p-column-filter"
-          placeholder="Search By Schema ID"
-          @input="filterCallback()"
-        />
-      </template>
-    </Column>
-    <Column :sortable="true" field="support_revocation" header="Revokable">
-      <template #body="{ data }">
-        <span v-if="data.support_revocation">
-          <i class="pi pi-check-circle"></i>
-        </span>
-      </template>
-    </Column>
-    <Column
-      :sortable="true"
-      field="created"
-      header="Created at"
-      filter-field="created"
-      :show-filter-match-modes="false"
-    >
-      <template #body="{ data }">
-        {{ data.created }}
-      </template>
-      <template #filter="{ filterModel, filterCallback }">
-        <InputText
-          v-model="filterModel.value"
-          type="text"
-          class="p-column-filter"
-          placeholder="Search By Time"
-          @input="filterCallback()"
-        />
-      </template>
-    </Column>
-    <template #expansion="{ data }">
-      <RowExpandData
-        :id="data.cred_def_id"
-        :url="API_PATH.CREDENTIAL_DEFINITION_STORAGE"
-      />
-    </template>
-  </DataTable>
+    </DataTable>
+  </MainCardContent>
 </template>
 
 <script setup lang="ts">
@@ -126,7 +121,7 @@ import { onMounted, ref, Ref, computed } from 'vue';
 // PrimeVue etc
 import Button from 'primevue/button';
 import Column from 'primevue/column';
-import DataTable, { DataTableFilterMetaData } from 'primevue/datatable';
+import DataTable from 'primevue/datatable';
 import InputText from 'primevue/inputtext';
 import { FilterMatchMode } from 'primevue/api';
 import { useConfirm } from 'primevue/useconfirm';
@@ -138,6 +133,7 @@ import { storeToRefs } from 'pinia';
 import RowExpandData from '../common/RowExpandData.vue';
 import { TABLE_OPT, API_PATH } from '@/helpers/constants';
 import { formatDateLong } from '@/helpers';
+import MainCardContent from '../layout/mainCard/MainCardContent.vue';
 
 const confirm = useConfirm();
 const toast = useToast();
@@ -202,19 +198,19 @@ const filter = ref({
   global: {
     value: null,
     matchMode: FilterMatchMode.CONTAINS,
-  } as DataTableFilterMetaData,
+  },
   cred_def_id: {
     value: null,
     matchMode: FilterMatchMode.CONTAINS,
-  } as DataTableFilterMetaData,
+  },
   schema_id: {
     value: null,
     matchMode: FilterMatchMode.CONTAINS,
-  } as DataTableFilterMetaData,
+  },
   created: {
     value: null,
     matchMode: FilterMatchMode.CONTAINS,
-  } as DataTableFilterMetaData,
+  },
 });
 </script>
 
@@ -222,18 +218,5 @@ const filter = ref({
 .row.buttons {
   float: right;
   margin: 3rem 1rem 0 0;
-}
-
-.p-datatable-header input {
-  padding-left: 3rem;
-  margin-right: 1rem;
-}
-
-.create-btn {
-  margin-right: 1rem;
-}
-
-.schema-search {
-  margin-left: 1.5rem;
 }
 </style>
