+++ services/tenant-ui/frontend/src/components/issuance/IssuedCredentials.vue
@@ -1,142 +1,139 @@
 <template>
-  <h3 class="mt-0">{{ $t('issue.credentials') }}</h3>
-
-  <DataTable
-    v-model:selection="selectedCredential"
-    v-model:filters="filter"
-    v-model:expandedRows="expandedRows"
-    :loading="loading"
-    :value="formattedCredentials"
-    :paginator="true"
-    :rows="TABLE_OPT.ROWS_DEFAULT"
-    :rows-per-page-options="TABLE_OPT.ROWS_OPTIONS"
-    selection-mode="single"
-    data-key="credential_exchange_id"
-    sort-field="created_at"
-    :sort-order="-1"
-    filter-display="menu"
+  <MainCardContent
+    :title="$t('issue.credentials')"
+    :refresh-callback="loadTable"
   >
-    <template #header>
-      <div class="flex justify-content-between">
-        <div class="flex justify-content-start">
-          <OfferCredential />
+    <DataTable
+      v-model:selection="selectedCredential"
+      v-model:filters="filter"
+      v-model:expandedRows="expandedRows"
+      :loading="loading"
+      :value="formattedCredentials"
+      :paginator="true"
+      :rows="TABLE_OPT.ROWS_DEFAULT"
+      :rows-per-page-options="TABLE_OPT.ROWS_OPTIONS"
+      selection-mode="single"
+      data-key="credential_exchange_id"
+      sort-field="created_at"
+      :sort-order="-1"
+      filter-display="menu"
+    >
+      <template #header>
+        <div class="flex justify-content-between">
+          <div class="flex justify-content-start">
+            <OfferCredential />
+          </div>
+          <div class="flex justify-content-end">
+            <span class="p-input-icon-left">
+              <i class="pi pi-search" />
+              <InputText
+                v-model="filter.global.value"
+                placeholder="Search Credentials"
+              />
+            </span>
+          </div>
         </div>
-        <div class="flex justify-content-end">
-          <span class="p-input-icon-left credential-search">
-            <i class="pi pi-search" />
-            <InputText
-              v-model="filter.global.value"
-              placeholder="Search Credentials"
-            />
-          </span>
-          <Button
-            icon="pi pi-refresh"
-            class="p-button-rounded p-button-outlined"
-            title="Refresh Table"
-            @click="loadTable"
-          ></Button>
-        </div>
-      </div>
-    </template>
-    <template #empty>{{ $t('common.noRecordsFound') }}</template>
-    <template #loading>{{ $t('common.loading') }}</template>
-    <Column :expander="true" header-style="width: 3rem" />
-    <Column header="Actions">
-      <template #body="{ data }">
-        <DeleteCredentialExchangeButton
-          :cred-exch-id="data.credential_exchange_id"
-        />
+      </template>
+      <template #empty>{{ $t('common.noRecordsFound') }}</template>
+      <template #loading>{{ $t('common.loading') }}</template>
+      <Column :expander="true" header-style="width: 3rem" />
+      <Column header="Actions">
+        <template #body="{ data }">
+          <DeleteCredentialExchangeButton
+            :cred-exch-id="data.credential_exchange_id"
+          />
 
-        <RevokeCredentialButton
-          :cred-exch-record="data"
-          :connection-display="findConnectionName(data.connection_id)"
+          <RevokeCredentialButton
+            :cred-exch-record="data"
+            :connection-display="findConnectionName(data.connection_id)"
+          />
+        </template>
+      </Column>
+      <Column
+        :sortable="true"
+        field="credential_definition_id"
+        header="Credential Definition"
+        filter-field="credential_definition_id"
+        :show-filter-match-modes="false"
+      >
+        <template #filter="{ filterModel, filterCallback }">
+          <InputText
+            v-model="filterModel.value"
+            type="text"
+            class="p-column-filter"
+            placeholder="Search By Credential Definition"
+            @input="filterCallback()"
+          />
+        </template>
+      </Column>
+      <Column
+        :sortable="true"
+        field="contact"
+        header="Contact"
+        filter-field="contact"
+        :show-filter-match-modes="false"
+      >
+        <template #body="{ data }">
+          {{ data.contact }}
+        </template>
+        <template #filter="{ filterModel, filterCallback }">
+          <InputText
+            v-model="filterModel.value"
+            type="text"
+            class="p-column-filter"
+            placeholder="Search By Contact"
+            @input="filterCallback()"
+          />
+        </template>
+      </Column>
+      <Column
+        :sortable="true"
+        field="state"
+        header="Status"
+        filter-field="state"
+        :show-filter-match-modes="false"
+      >
+        <template #body="{ data }">
+          <StatusChip :status="data.state" />
+        </template>
+        <template #filter="{ filterModel, filterCallback }">
+          <InputText
+            v-model="filterModel.value"
+            type="text"
+            class="p-column-filter"
+            placeholder="Search By Status"
+            @input="filterCallback()"
+          />
+        </template>
+      </Column>
+      <Column
+        :sortable="true"
+        field="created"
+        header="Created at"
+        filter-field="created"
+        :show-filter-match-modes="false"
+      >
+        <template #body="{ data }">
+          {{ data.created }}
+        </template>
+        <template #filter="{ filterModel, filterCallback }">
+          <InputText
+            v-model="filterModel.value"
+            type="text"
+            class="p-column-filter"
+            placeholder="Search By Time"
+            @input="filterCallback()"
+          />
+        </template>
+      </Column>
+      <template #expansion="{ data }">
+        <RowExpandData
+          :id="data.credential_exchange_id"
+          :url="API_PATH.ISSUE_CREDENTIAL_RECORDS"
         />
       </template>
-    </Column>
-    <Column
-      :sortable="true"
-      field="credential_definition_id"
-      header="Credential Definition"
-      filter-field="credential_definition_id"
-      :show-filter-match-modes="false"
-    >
-      <template #filter="{ filterModel, filterCallback }">
-        <InputText
-          v-model="filterModel.value"
-          type="text"
-          class="p-column-filter"
-          placeholder="Search By Credential Definition"
-          @input="filterCallback()"
-        />
-      </template>
-    </Column>
-    <Column
-      :sortable="true"
-      field="contact"
-      header="Contact"
-      filter-field="contact"
-      :show-filter-match-modes="false"
-    >
-      <template #body="{ data }">
-        {{ data.contact }}
-      </template>
-      <template #filter="{ filterModel, filterCallback }">
-        <InputText
-          v-model="filterModel.value"
-          type="text"
-          class="p-column-filter"
-          placeholder="Search By Contact"
-          @input="filterCallback()"
-        />
-      </template>
-    </Column>
-    <Column
-      :sortable="true"
-      field="state"
-      header="Status"
-      filter-field="state"
-      :show-filter-match-modes="false"
-    >
-      <template #body="{ data }">
-        <StatusChip :status="data.state" />
-      </template>
-      <template #filter="{ filterModel, filterCallback }">
-        <InputText
-          v-model="filterModel.value"
-          type="text"
-          class="p-column-filter"
-          placeholder="Search By Status"
-          @input="filterCallback()"
-        />
-      </template>
-    </Column>
-    <Column
-      :sortable="true"
-      field="created"
-      header="Created at"
-      filter-field="created"
-      :show-filter-match-modes="false"
-    >
-      <template #body="{ data }">
-        {{ data.created }}
-      </template>
-      <template #filter="{ filterModel, filterCallback }">
-        <InputText
-          v-model="filterModel.value"
-          type="text"
-          class="p-column-filter"
-          placeholder="Search By Time"
-          @input="filterCallback()"
-        />
-      </template>
-    </Column>
-    <template #expansion="{ data }">
-      <RowExpandData
-        :id="data.credential_exchange_id"
-        :url="API_PATH.ISSUE_CREDENTIAL_RECORDS"
-      />
-    </template>
-  </DataTable>
+    </DataTable>
+  </MainCardContent>
 </template>
 
 <script setup lang="ts">
@@ -146,9 +143,8 @@ import { onMounted, ref, Ref, computed } from 'vue';
 import { useIssuerStore, useContactsStore } from '@/store';
 import { storeToRefs } from 'pinia';
 // PrimeVue/etc
-import Button from 'primevue/button';
 import Column from 'primevue/column';
-import DataTable, { DataTableFilterMetaData } from 'primevue/datatable';
+import DataTable from 'primevue/datatable';
 import InputText from 'primevue/inputtext';
 import { FilterMatchMode } from 'primevue/api';
 import { useToast } from 'vue-toastification';
@@ -156,6 +152,7 @@ import { useToast } from 'vue-toastification';
 import { TABLE_OPT, API_PATH } from '@/helpers/constants';
 import { formatDateLong } from '@/helpers';
 import DeleteCredentialExchangeButton from './deleteCredential/DeleteCredentialExchangeButton.vue';
+import MainCardContent from '../layout/mainCard/MainCardContent.vue';
 import OfferCredential from './offerCredential/OfferCredential.vue';
 import RevokeCredentialButton from './deleteCredential/RevokeCredentialButton.vue';
 import RowExpandData from '../common/RowExpandData.vue';
@@ -183,6 +180,7 @@ const formattedCredentials: Ref<any[]> = computed(() =>
     state: cred.state,
     contact: findConnectionName(cred.connection_id),
     credential_definition_id: cred.credential_definition_id,
+    credential_exchange_id: cred.credential_exchange_id,
     sent_time: cred.sent_time,
     created: formatDateLong(cred.created_at),
     created_at: cred.created_at,
@@ -217,28 +215,18 @@ const filter = ref({
   global: {
     value: null,
     matchMode: FilterMatchMode.CONTAINS,
-  } as DataTableFilterMetaData,
+  },
   contact: {
     value: null,
     matchMode: FilterMatchMode.CONTAINS,
-  } as DataTableFilterMetaData,
+  },
   credential_definition_id: {
     value: null,
     matchMode: FilterMatchMode.CONTAINS,
-  } as DataTableFilterMetaData,
+  },
   created: {
     value: null,
     matchMode: FilterMatchMode.CONTAINS,
-  } as DataTableFilterMetaData,
+  },
 });
 </script>
-
-<style scoped>
-.credential-search {
-  margin-left: 1.5rem;
-}
-.credential-search input {
-  padding-left: 3rem !important;
-  margin-right: 1rem;
-}
-</style>
