+++ services/tenant-ui/frontend/src/components/oca/Oca.vue
@@ -1,116 +1,111 @@
 <template>
-  <h3 class="mt-0">
-    {{ $t('configuration.oca.oca') }}
-  </h3>
-
-  <DataTable
-    v-model:expandedRows="expandedRows"
-    v-model:filters="filter"
-    :loading="loading"
-    :value="formattedOcas"
-    :paginator="true"
-    :rows="TABLE_OPT.ROWS_DEFAULT"
-    :rows-per-page-options="TABLE_OPT.ROWS_OPTIONS"
-    selection-mode="single"
-    data-key="oca_id"
-    sort-field="created_at"
-    :sort-order="-1"
-    filter-display="menu"
+  <MainCardContent
+    :title="$t('configuration.oca.oca')"
+    :refresh-callback="loadTable"
   >
-    <template #header>
-      <div class="flex justify-content-between">
-        <div class="flex justify-content-start">
-          <CreateOca />
+    <DataTable
+      v-model:expandedRows="expandedRows"
+      v-model:filters="filter"
+      :loading="loading"
+      :value="formattedOcas"
+      :paginator="true"
+      :rows="TABLE_OPT.ROWS_DEFAULT"
+      :rows-per-page-options="TABLE_OPT.ROWS_OPTIONS"
+      selection-mode="single"
+      data-key="oca_id"
+      sort-field="created_at"
+      :sort-order="-1"
+      filter-display="menu"
+    >
+      <template #header>
+        <div class="flex justify-content-between">
+          <div class="flex justify-content-start">
+            <CreateOca />
+          </div>
+          <div class="flex justify-content-end">
+            <span class="p-input-icon-left">
+              <i class="pi pi-search" />
+              <InputText
+                v-model="filter.global.value"
+                placeholder="Search OCA Items"
+              />
+            </span>
+          </div>
         </div>
-        <div class="flex justify-content-end">
-          <span class="p-input-icon-left oca-search">
-            <i class="pi pi-search" />
-            <InputText
-              v-model="filter.global.value"
-              placeholder="Search OCA Items"
-            />
-          </span>
+      </template>
+      <template #empty>{{ $t('common.noRecordsFound') }}</template>
+      <template #loading>{{ $t('common.loading') }}</template>
+      <Column :expander="true" header-style="width: 3rem" />
+      <Column :sortable="false" header="Actions">
+        <template #body="{ data }">
           <Button
-            icon="pi pi-refresh"
-            class="p-button-rounded p-button-outlined"
-            title="Refresh Table"
-            @click="loadTable"
+            title="Delete Credential Definition"
+            icon="pi pi-trash"
+            class="p-button-rounded p-button-icon-only p-button-text"
+            @click="deleteOca($event, data.oca_id)"
           />
-        </div>
-      </div>
-    </template>
-    <template #empty>{{ $t('common.noRecordsFound') }}</template>
-    <template #loading>{{ $t('common.loading') }}</template>
-    <Column :expander="true" header-style="width: 3rem" />
-    <Column :sortable="false" header="Actions">
-      <template #body="{ data }">
-        <Button
-          title="Delete Credential Definition"
-          icon="pi pi-trash"
-          class="p-button-rounded p-button-icon-only p-button-text"
-          @click="deleteOca($event, data.oca_id)"
-        />
-      </template>
-    </Column>
-    <Column
-      :sortable="true"
-      field="cred_def_id"
-      header="Cred Def ID"
-      filter-field="cred_def_id"
-      :show-filter-match-modes="false"
-    >
-      <template #filter="{ filterModel, filterCallback }">
-        <InputText
-          v-model="filterModel.value"
-          type="text"
-          class="p-column-filter"
-          placeholder="Search By Cred Def ID"
-          @input="filterCallback()"
-        />
-      </template>
-    </Column>
+        </template>
+      </Column>
+      <Column
+        :sortable="true"
+        field="cred_def_id"
+        header="Cred Def ID"
+        filter-field="cred_def_id"
+        :show-filter-match-modes="false"
+      >
+        <template #filter="{ filterModel, filterCallback }">
+          <InputText
+            v-model="filterModel.value"
+            type="text"
+            class="p-column-filter"
+            placeholder="Search By Cred Def ID"
+            @input="filterCallback()"
+          />
+        </template>
+      </Column>
 
-    <Column :sortable="true" header="OCA Bundle">
-      <template #body="{ data }">
-        <span v-if="data.bundle">
-          <i
-            v-tooltip="'Bundle JSON stored in Traction, expand row to view'"
-            class="pi pi-database"
-          >
-          </i>
-          {{ $t('oca.json') }}
-        </span>
-        <span v-else-if="data.url">
-          <i v-tooltip="'Bundle URL'" class="pi pi-link"> </i>
-          {{ data.url }}
-        </span>
-      </template>
-    </Column>
+      <Column :sortable="true" header="OCA Bundle">
+        <template #body="{ data }">
+          <span v-if="data.bundle">
+            <i
+              v-tooltip="'Bundle JSON stored in Traction, expand row to view'"
+              class="pi pi-database"
+            >
+            </i>
+            {{ $t('oca.json') }}
+          </span>
+          <span v-else-if="data.url">
+            <i v-tooltip="'Bundle URL'" class="pi pi-link"> </i>
+            {{ data.url }}
+          </span>
+        </template>
+      </Column>
 
-    <Column
-      :sortable="true"
-      field="created"
-      header="Created at"
-      filter-field="created"
-      :show-filter-match-modes="false"
-    >
-      <template #body="{ data }">
-        {{ data.created }}
+      <Column
+        :sortable="true"
+        field="created"
+        header="Created at"
+        filter-field="created"
+        :show-filter-match-modes="false"
+      >
+        <template #body="{ data }">
+          {{ data.created }}
+        </template>
+        <template #filter="{ filterModel, filterCallback }">
+          <InputText
+            v-model="filterModel.value"
+            type="text"
+            class="p-column-filter"
+            placeholder="Search By Time"
+            @input="filterCallback()"
+          />
+        </template>
+      </Column>
+      <template #expansion="{ data }">
+        <RowExpandData :id="data.oca_id" :url="API_PATH.OCAS" />
       </template>
-      <template #filter="{ filterModel, filterCallback }">
-        <InputText
-          v-model="filterModel.value"
-          type="text"
-          class="p-column-filter"
-          placeholder="Search By Time"
-          @input="filterCallback()"
-        />
-      </template>
-    </Column>
-    <template #expansion="{ data }">
-      <RowExpandData :id="data.oca_id" :url="API_PATH.OCAS" />
-    </template>
-  </DataTable>
+    </DataTable>
+  </MainCardContent>
 </template>
 
 <script setup lang="ts">
@@ -119,7 +114,7 @@ import { onMounted, ref, Ref, computed } from 'vue';
 // PrimeVue etc
 import Button from 'primevue/button';
 import Column from 'primevue/column';
-import DataTable, { DataTableFilterMetaData } from 'primevue/datatable';
+import DataTable from 'primevue/datatable';
 import InputText from 'primevue/inputtext';
 import { FilterMatchMode } from 'primevue/api';
 import { useConfirm } from 'primevue/useconfirm';
@@ -129,6 +124,7 @@ import { useGovernanceStore } from '../../store';
 import { storeToRefs } from 'pinia';
 // Custom components
 import CreateOca from './createOca/CreateOca.vue';
+import MainCardContent from '../layout/mainCard/MainCardContent.vue';
 import RowExpandData from '../common/RowExpandData.vue';
 import { TABLE_OPT, API_PATH } from '@/helpers/constants';
 import { formatDateLong } from '@/helpers';
@@ -195,15 +191,15 @@ const filter = ref({
   global: {
     value: null,
     matchMode: FilterMatchMode.CONTAINS,
-  } as DataTableFilterMetaData,
+  },
   created: {
     value: null,
     matchMode: FilterMatchMode.CONTAINS,
-  } as DataTableFilterMetaData,
+  },
   cred_def_id: {
     value: null,
     matchMode: FilterMatchMode.CONTAINS,
-  } as DataTableFilterMetaData,
+  },
 });
 </script>
 
@@ -211,18 +207,5 @@ const filter = ref({
 .row.buttons {
   float: right;
   margin: 3rem 1rem 0 0;
-}
-
-.p-datatable-header input {
-  padding-left: 3rem;
-  margin-right: 1rem;
-}
-
-.create-btn {
-  margin-right: 1rem;
-}
-
-.oca-search {
-  margin-left: 1.5rem;
 }
 </style>
