+++ services/tenant-ui/frontend/src/components/profile/SettingsForm.vue
@@ -1,332 +1,340 @@
 <template>
-  <form @submit.prevent="handleSubmit(!v$.$invalid)">
-    <ProgressSpinner v-if="loading" />
-    <div v-else class="settings-form">
-      <!-- Wallet Label -->
-      <div class="field">
-        <label for="walletLabel">{{ $t('profile.walletLabel') }}</label>
-        <InputText
-          id="walletLabel"
-          v-model="v$.walletLabel.$model"
-          class="w-full"
-          :class="{ 'p-invalid': v$.walletLabel.$invalid && submitted }"
-        />
-        <span v-if="v$.walletLabel.$error && submitted">
-          <span v-for="(error, index) of v$.walletLabel.$errors" :key="index">
-            <small class="p-error">{{ error.$message }}</small>
+  <MainCardContent
+    :title="$t('tenant.profile.tenantProfile')"
+    :refresh-callback="loadTenantSettings"
+  >
+    <form @submit.prevent="handleSubmit(!v$.$invalid)">
+      <ProgressSpinner v-if="loading" />
+      <div v-else class="settings-form">
+        <!-- Wallet Label -->
+        <div class="field">
+          <label for="walletLabel">{{ $t('profile.walletLabel') }}</label>
+          <InputText
+            id="walletLabel"
+            v-model="v$.walletLabel.$model"
+            class="w-full"
+            :class="{ 'p-invalid': v$.walletLabel.$invalid && submitted }"
+          />
+          <span v-if="v$.walletLabel.$error && submitted">
+            <span v-for="(error, index) of v$.walletLabel.$errors" :key="index">
+              <small class="p-error">{{ error.$message }}</small>
+            </span>
           </span>
-        </span>
-      </div>
+        </div>
 
-      <!-- Webhooks -->
-      <div class="webhooks">
-        <TransitionGroup appear name="wh">
-          <div
-            v-for="(webhook, index) of formFields.webhooks"
-            :key="index"
-            class="webhook"
-          >
-            <div class="field">
-              <label for="webhookUrl">{{ $t('profile.webHookUrl') }}</label>
-              <InputText
-                id="webhookUrl"
-                v-model="webhook.webhookUrl"
-                class="w-full"
+        <!-- Webhooks -->
+        <div class="webhooks">
+          <TransitionGroup appear name="wh">
+            <div
+              v-for="(webhook, index) of formFields.webhooks"
+              :key="index"
+              class="webhook"
+            >
+              <div class="field">
+                <label for="webhookUrl">{{ $t('profile.webHookUrl') }}</label>
+                <InputText
+                  id="webhookUrl"
+                  v-model="webhook.webhookUrl"
+                  class="w-full"
+                />
+              </div>
+
+              <div class="field">
+                <label for="webhookKey">{{ $t('profile.webHookKey') }}</label>
+                <Password
+                  id="webhookKey"
+                  v-model="webhook.webhookKey"
+                  class="w-full"
+                  toggle-mask
+                  :feedback="false"
+                />
+              </div>
+
+              <Button
+                title="Delete this webhook"
+                icon="pi pi-trash"
+                text
+                rounded
+                @click="() => removeWebhook(index)"
+              />
+
+              <Button
+                title="Add another webhook"
+                class="add"
+                icon="pi pi-plus-circle"
+                text
+                rounded
+                @click="addWebhook"
               />
             </div>
+          </TransitionGroup>
+        </div>
 
-            <div class="field">
-              <label for="webhookKey">{{ $t('profile.webHookKey') }}</label>
-              <Password
-                id="webhookKey"
-                v-model="webhook.webhookKey"
-                class="w-full"
-                toggle-mask
-                :feedback="false"
+        <Divider class="mt-2" />
+
+        <!-- Image URL -->
+        <div class="field">
+          <label for="imageUrl">{{ $t('profile.imageUrl') }}</label>
+          <InputText
+            id="imageUrl"
+            v-model="v$.imageUrl.$model"
+            class="w-full"
+            :class="{ 'p-invalid': v$.imageUrl.$invalid && submitted }"
+          />
+          <span v-if="v$.imageUrl.$error && submitted">
+            <span v-for="(error, index) of v$.imageUrl.$errors" :key="index">
+              <small class="p-error">{{ error.$message }}</small>
+            </span>
+          </span>
+        </div>
+
+        <!-- Extra Acapy Settings -->
+        <Panel v-if="!hideAcapySettings" class="settings-group mb-5" toggleable>
+          <template #header>
+            <b>{{ $t('tenant.settings.extraSettings') }}</b>
+          </template>
+          <div class="field">
+            <label for="ACAPY_ENDORSER_ROLE">
+              {{ $t('tenant.settings.acapyEndorserRole') }}
+              <i
+                v-tooltip="$t('tenant.settings.acapyHelpEndorserRole')"
+                class="pi pi-question-circle"
               />
-            </div>
-
-            <Button
-              title="Delete this webhook"
-              icon="pi pi-trash"
-              text
-              rounded
-              @click="() => removeWebhook(index)"
-            />
-
-            <Button
-              title="Add another webhook"
-              class="add"
-              icon="pi pi-plus-circle"
-              text
-              rounded
-              @click="addWebhook"
+            </label>
+            <Dropdown
+              v-model="v$.ACAPY_ENDORSER_ROLE.$model"
+              :options="endorserRole"
+              class="w-full"
             />
           </div>
-        </TransitionGroup>
-      </div>
 
-      <Divider class="mt-2" />
+          <div class="field">
+            <label for="ACAPY_INVITE_PUBLIC">
+              {{ $t('tenant.settings.acapyInvitePublic') }}
+              <i
+                v-tooltip="$t('tenant.settings.acapyHelpInvitePublic')"
+                class="pi pi-question-circle"
+              />
+            </label>
+            <InputSwitch
+              id="ACAPY_INVITE_PUBLIC"
+              v-model="v$.ACAPY_INVITE_PUBLIC.$model"
+            />
+          </div>
 
-      <!-- Image URL -->
-      <div class="field">
-        <label for="imageUrl">{{ $t('profile.imageUrl') }}</label>
-        <InputText
-          id="imageUrl"
-          v-model="v$.imageUrl.$model"
-          class="w-full"
-          :class="{ 'p-invalid': v$.imageUrl.$invalid && submitted }"
+          <div class="field">
+            <label for="ACAPY_LOG_LEVEL">
+              {{ $t('tenant.settings.acapyLogLevel') }}
+              <i
+                v-tooltip="$t('tenant.settings.acapyHelpLogLevel')"
+                class="pi pi-question-circle"
+              />
+            </label>
+            <Dropdown
+              v-model="v$.ACAPY_LOG_LEVEL.$model"
+              :options="logLevels"
+              class="w-full"
+            />
+          </div>
+
+          <div class="field">
+            <label for="ACAPY_MONITOR_PING">
+              {{ $t('tenant.settings.acapyMonitorPing') }}
+              <i
+                v-tooltip="$t('tenant.settings.acapyHelpMonitorPing')"
+                class="pi pi-question-circle"
+              />
+            </label>
+            <InputSwitch
+              id="ACAPY_MONITOR_PING"
+              v-model="v$.ACAPY_MONITOR_PING.$model"
+            />
+          </div>
+
+          <div class="field">
+            <label for="ACAPY_NOTIFY_REVOCATION">
+              {{ $t('tenant.settings.acapyNotifyRevocation') }}
+              <i
+                v-tooltip="$t('tenant.settings.acapyHelpNotifyRevocation')"
+                class="pi pi-question-circle"
+              />
+            </label>
+            <InputSwitch
+              id="ACAPY_NOTIFY_REVOCATION"
+              v-model="v$.ACAPY_NOTIFY_REVOCATION.$model"
+            />
+          </div>
+
+          <div class="field">
+            <label for="ACAPY_PUBLIC_INVITES">
+              {{ $t('tenant.settings.acapyPublicInvites') }}
+              <i
+                v-tooltip="$t('tenant.settings.acapyHelpPublicInvites')"
+                class="pi pi-question-circle"
+              />
+            </label>
+            <InputSwitch
+              id="ACAPY_PUBLIC_INVITES"
+              v-model="v$.ACAPY_PUBLIC_INVITES.$model"
+            />
+          </div>
+
+          <p>
+            <strong>{{ $t('tenant.settings.auto') }}</strong>
+          </p>
+
+          <div class="field">
+            <label for="ACAPY_AUTO_ACCEPT_INVITES">
+              {{ $t('tenant.settings.acapyAutoAcceptInvites') }}
+              <i
+                v-tooltip="$t('tenant.settings.acapyHelpAutoAcceptInvites')"
+                class="pi pi-question-circle"
+              />
+            </label>
+            <InputSwitch
+              id="ACAPY_AUTO_ACCEPT_INVITES"
+              v-model="v$.ACAPY_AUTO_ACCEPT_INVITES.$model"
+            />
+          </div>
+
+          <div class="field">
+            <label for="ACAPY_AUTO_ACCEPT_REQUESTS">
+              {{ $t('tenant.settings.acapyAutoAcceptRequests') }}
+              <i
+                v-tooltip="$t('tenant.settings.acapyHelpAutoAcceptRequests')"
+                class="pi pi-question-circle"
+              />
+            </label>
+            <InputSwitch
+              id="ACAPY_AUTO_ACCEPT_REQUESTS"
+              v-model="v$.ACAPY_AUTO_ACCEPT_REQUESTS.$model"
+            />
+          </div>
+
+          <div class="field">
+            <label for="ACAPY_AUTO_PING_CONNECTION">
+              {{ $t('tenant.settings.acapyAutoPingConnection') }}
+              <i
+                v-tooltip="$t('tenant.settings.acapyHelpAutoPingConnection')"
+                class="pi pi-question-circle"
+              />
+            </label>
+            <InputSwitch
+              id="ACAPY_AUTO_PING_CONNECTION"
+              v-model="v$.ACAPY_AUTO_PING_CONNECTION.$model"
+            />
+          </div>
+
+          <div class="field">
+            <label for="ACAPY_AUTO_REQUEST_ENDORSEMENT">
+              {{ $t('tenant.settings.acapyAutoRequestEndorsement') }}
+              <i
+                v-tooltip="
+                  $t('tenant.settings.acapyHelpAutoRequestEndorsement')
+                "
+                class="pi pi-question-circle"
+              />
+            </label>
+            <InputSwitch
+              id=" ACAPY_AUTO_REQUEST_ENDORSEMENT "
+              v-model="v$.ACAPY_AUTO_REQUEST_ENDORSEMENT.$model"
+            />
+          </div>
+
+          <div class="field">
+            <label for="ACAPY_AUTO_RESPOND_CREDENTIAL_OFFER">
+              {{ $t('tenant.settings.acapyAutoRespondCredentialOffer') }}
+              <i
+                v-tooltip="
+                  $t('tenant.settings.acapyHelpAutoRespondCredentialOffer')
+                "
+                class="pi pi-question-circle"
+              />
+            </label>
+            <InputSwitch
+              id="ACAPY_AUTO_RESPOND_CREDENTIAL_OFFER"
+              v-model="v$.ACAPY_AUTO_RESPOND_CREDENTIAL_OFFER.$model"
+            />
+          </div>
+
+          <div class="field">
+            <label for="ACAPY_AUTO_RESPOND_CREDENTIAL_REQUEST">
+              {{ $t('tenant.settings.acapyAutoRespondCredentialRequest') }}
+              <i
+                v-tooltip="
+                  $t('tenant.settings.acapyHelpAutoRespondCredentialRequest')
+                "
+                class="pi pi-question-circle"
+              />
+            </label>
+            <InputSwitch
+              id="ACAPY_AUTO_RESPOND_CREDENTIAL_REQUEST"
+              v-model="v$.ACAPY_AUTO_RESPOND_CREDENTIAL_REQUEST.$model"
+            />
+          </div>
+
+          <div class="field">
+            <label for="ACAPY_AUTO_RESPOND_MESSAGES">
+              {{ $t('tenant.settings.acapyAutoRespondMessages') }}
+              <i
+                v-tooltip="$t('tenant.settings.acapyHelpAutoRespondMessages')"
+                class="pi pi-question-circle"
+              />
+            </label>
+            <InputSwitch
+              id="ACAPY_AUTO_RESPOND_MESSAGES"
+              v-model="v$.ACAPY_AUTO_RESPOND_MESSAGES.$model"
+            />
+          </div>
+
+          <div class="field">
+            <label for="ACAPY_AUTO_WRITE_TRANSACTIONS">
+              {{ $t('tenant.settings.acapyAutoWriteTransactions') }}
+              <i
+                v-tooltip="$t('tenant.settings.acapyHelpAutoWriteTransactions')"
+                class="pi pi-question-circle"
+              />
+            </label>
+            <InputSwitch
+              id=" ACAPY_AUTO_WRITE_TRANSACTIONS "
+              v-model="v$.ACAPY_AUTO_WRITE_TRANSACTIONS.$model"
+            />
+          </div>
+
+          <div class="field">
+            <label for="ACAPY_AUTO_VERIFY_PRESENTATION">
+              {{ $t('tenant.settings.acapyAutoVerifyPresentation') }}
+              <i
+                v-tooltip="
+                  $t('tenant.settings.acapyHelpAutoVerifyPresentation')
+                "
+                class="pi pi-question-circle"
+              />
+            </label>
+            <InputSwitch
+              id="ACAPY_AUTO_VERIFY_PRESENTATION"
+              v-model="v$.ACAPY_AUTO_VERIFY_PRESENTATION.$model"
+            />
+          </div>
+        </Panel>
+
+        <div>
+          <Accordion>
+            <AccordionTab header="Tenant Wallet Details">
+              <vue-json-pretty :data="tenantWallet" />
+            </AccordionTab>
+          </Accordion>
+        </div>
+        <Button
+          class="mt-6 mb-3"
+          :disabled="loading"
+          :loading="loading"
+          label="Save Changes"
+          type="submit"
         />
-        <span v-if="v$.imageUrl.$error && submitted">
-          <span v-for="(error, index) of v$.imageUrl.$errors" :key="index">
-            <small class="p-error">{{ error.$message }}</small>
-          </span>
-        </span>
       </div>
-
-      <!-- Extra Acapy Settings -->
-      <Panel v-if="!hideAcapySettings" class="settings-group mb-5" toggleable>
-        <template #header>
-          <b>{{ $t('tenant.settings.extraSettings') }}</b>
-        </template>
-        <div class="field">
-          <label for="ACAPY_ENDORSER_ROLE">
-            {{ $t('tenant.settings.acapyEndorserRole') }}
-            <i
-              v-tooltip="$t('tenant.settings.acapyHelpEndorserRole')"
-              class="pi pi-question-circle"
-            />
-          </label>
-          <Dropdown
-            v-model="v$.ACAPY_ENDORSER_ROLE.$model"
-            :options="endorserRole"
-            class="w-full"
-          />
-        </div>
-
-        <div class="field">
-          <label for="ACAPY_INVITE_PUBLIC">
-            {{ $t('tenant.settings.acapyInvitePublic') }}
-            <i
-              v-tooltip="$t('tenant.settings.acapyHelpInvitePublic')"
-              class="pi pi-question-circle"
-            />
-          </label>
-          <InputSwitch
-            id="ACAPY_INVITE_PUBLIC"
-            v-model="v$.ACAPY_INVITE_PUBLIC.$model"
-          />
-        </div>
-
-        <div class="field">
-          <label for="ACAPY_LOG_LEVEL">
-            {{ $t('tenant.settings.acapyLogLevel') }}
-            <i
-              v-tooltip="$t('tenant.settings.acapyHelpLogLevel')"
-              class="pi pi-question-circle"
-            />
-          </label>
-          <Dropdown
-            v-model="v$.ACAPY_LOG_LEVEL.$model"
-            :options="logLevels"
-            class="w-full"
-          />
-        </div>
-
-        <div class="field">
-          <label for="ACAPY_MONITOR_PING">
-            {{ $t('tenant.settings.acapyMonitorPing') }}
-            <i
-              v-tooltip="$t('tenant.settings.acapyHelpMonitorPing')"
-              class="pi pi-question-circle"
-            />
-          </label>
-          <InputSwitch
-            id="ACAPY_MONITOR_PING"
-            v-model="v$.ACAPY_MONITOR_PING.$model"
-          />
-        </div>
-
-        <div class="field">
-          <label for="ACAPY_NOTIFY_REVOCATION">
-            {{ $t('tenant.settings.acapyNotifyRevocation') }}
-            <i
-              v-tooltip="$t('tenant.settings.acapyHelpNotifyRevocation')"
-              class="pi pi-question-circle"
-            />
-          </label>
-          <InputSwitch
-            id="ACAPY_NOTIFY_REVOCATION"
-            v-model="v$.ACAPY_NOTIFY_REVOCATION.$model"
-          />
-        </div>
-
-        <div class="field">
-          <label for="ACAPY_PUBLIC_INVITES">
-            {{ $t('tenant.settings.acapyPublicInvites') }}
-            <i
-              v-tooltip="$t('tenant.settings.acapyHelpPublicInvites')"
-              class="pi pi-question-circle"
-            />
-          </label>
-          <InputSwitch
-            id="ACAPY_PUBLIC_INVITES"
-            v-model="v$.ACAPY_PUBLIC_INVITES.$model"
-          />
-        </div>
-
-        <p>
-          <strong>{{ $t('tenant.settings.auto') }}</strong>
-        </p>
-
-        <div class="field">
-          <label for="ACAPY_AUTO_ACCEPT_INVITES">
-            {{ $t('tenant.settings.acapyAutoAcceptInvites') }}
-            <i
-              v-tooltip="$t('tenant.settings.acapyHelpAutoAcceptInvites')"
-              class="pi pi-question-circle"
-            />
-          </label>
-          <InputSwitch
-            id="ACAPY_AUTO_ACCEPT_INVITES"
-            v-model="v$.ACAPY_AUTO_ACCEPT_INVITES.$model"
-          />
-        </div>
-
-        <div class="field">
-          <label for="ACAPY_AUTO_ACCEPT_REQUESTS">
-            {{ $t('tenant.settings.acapyAutoAcceptRequests') }}
-            <i
-              v-tooltip="$t('tenant.settings.acapyHelpAutoAcceptRequests')"
-              class="pi pi-question-circle"
-            />
-          </label>
-          <InputSwitch
-            id="ACAPY_AUTO_ACCEPT_REQUESTS"
-            v-model="v$.ACAPY_AUTO_ACCEPT_REQUESTS.$model"
-          />
-        </div>
-
-        <div class="field">
-          <label for="ACAPY_AUTO_PING_CONNECTION">
-            {{ $t('tenant.settings.acapyAutoPingConnection') }}
-            <i
-              v-tooltip="$t('tenant.settings.acapyHelpAutoPingConnection')"
-              class="pi pi-question-circle"
-            />
-          </label>
-          <InputSwitch
-            id="ACAPY_AUTO_PING_CONNECTION"
-            v-model="v$.ACAPY_AUTO_PING_CONNECTION.$model"
-          />
-        </div>
-
-        <div class="field">
-          <label for="ACAPY_AUTO_REQUEST_ENDORSEMENT">
-            {{ $t('tenant.settings.acapyAutoRequestEndorsement') }}
-            <i
-              v-tooltip="$t('tenant.settings.acapyHelpAutoRequestEndorsement')"
-              class="pi pi-question-circle"
-            />
-          </label>
-          <InputSwitch
-            id=" ACAPY_AUTO_REQUEST_ENDORSEMENT "
-            v-model="v$.ACAPY_AUTO_REQUEST_ENDORSEMENT.$model"
-          />
-        </div>
-
-        <div class="field">
-          <label for="ACAPY_AUTO_RESPOND_CREDENTIAL_OFFER">
-            {{ $t('tenant.settings.acapyAutoRespondCredentialOffer') }}
-            <i
-              v-tooltip="
-                $t('tenant.settings.acapyHelpAutoRespondCredentialOffer')
-              "
-              class="pi pi-question-circle"
-            />
-          </label>
-          <InputSwitch
-            id="ACAPY_AUTO_RESPOND_CREDENTIAL_OFFER"
-            v-model="v$.ACAPY_AUTO_RESPOND_CREDENTIAL_OFFER.$model"
-          />
-        </div>
-
-        <div class="field">
-          <label for="ACAPY_AUTO_RESPOND_CREDENTIAL_REQUEST">
-            {{ $t('tenant.settings.acapyAutoRespondCredentialRequest') }}
-            <i
-              v-tooltip="
-                $t('tenant.settings.acapyHelpAutoRespondCredentialRequest')
-              "
-              class="pi pi-question-circle"
-            />
-          </label>
-          <InputSwitch
-            id="ACAPY_AUTO_RESPOND_CREDENTIAL_REQUEST"
-            v-model="v$.ACAPY_AUTO_RESPOND_CREDENTIAL_REQUEST.$model"
-          />
-        </div>
-
-        <div class="field">
-          <label for="ACAPY_AUTO_RESPOND_MESSAGES">
-            {{ $t('tenant.settings.acapyAutoRespondMessages') }}
-            <i
-              v-tooltip="$t('tenant.settings.acapyHelpAutoRespondMessages')"
-              class="pi pi-question-circle"
-            />
-          </label>
-          <InputSwitch
-            id="ACAPY_AUTO_RESPOND_MESSAGES"
-            v-model="v$.ACAPY_AUTO_RESPOND_MESSAGES.$model"
-          />
-        </div>
-
-        <div class="field">
-          <label for="ACAPY_AUTO_WRITE_TRANSACTIONS">
-            {{ $t('tenant.settings.acapyAutoWriteTransactions') }}
-            <i
-              v-tooltip="$t('tenant.settings.acapyHelpAutoWriteTransactions')"
-              class="pi pi-question-circle"
-            />
-          </label>
-          <InputSwitch
-            id=" ACAPY_AUTO_WRITE_TRANSACTIONS "
-            v-model="v$.ACAPY_AUTO_WRITE_TRANSACTIONS.$model"
-          />
-        </div>
-
-        <div class="field">
-          <label for="ACAPY_AUTO_VERIFY_PRESENTATION">
-            {{ $t('tenant.settings.acapyAutoVerifyPresentation') }}
-            <i
-              v-tooltip="$t('tenant.settings.acapyHelpAutoVerifyPresentation')"
-              class="pi pi-question-circle"
-            />
-          </label>
-          <InputSwitch
-            id="ACAPY_AUTO_VERIFY_PRESENTATION"
-            v-model="v$.ACAPY_AUTO_VERIFY_PRESENTATION.$model"
-          />
-        </div>
-      </Panel>
-
-      <div>
-        <Accordion>
-          <AccordionTab header="Tenant Wallet Details">
-            <vue-json-pretty :data="tenantWallet" />
-          </AccordionTab>
-        </Accordion>
-      </div>
-    </div>
-
-    <Button
-      class="mt-6 mb-3"
-      :disabled="loading"
-      :loading="loading"
-      label="Save Changes"
-      type="submit"
-    />
-  </form>
+    </form>
+  </MainCardContent>
 </template>
 
 <script setup lang="ts">
@@ -350,6 +358,8 @@ import { useVuelidate } from '@vuelidate/core';
 // State/etc
 import { storeToRefs } from 'pinia';
 import { useTenantStore } from '@/store';
+// Components
+import MainCardContent from '@/components/layout/mainCard/MainCardContent.vue';
 
 const toast = useToast();
