+++ services/tenant-ui/frontend/src/plugins/i18n/locales/en.json
@@ -143,7 +143,9 @@
         "connection": "Connection",
         "credential": "Credential",
         "lastUpdate": "Last Update"
-      }
+      },
+      "viewCard": "View in Card format",
+      "viewTable": "View in Table format"
     },
     "wallet": {
       "notFound": "There are no credentials in your Wallet",
