+++ services/tenant-ui/frontend/src/views/connections/MyInvitations.vue
@@ -3,5 +3,5 @@
 </template>
 
 <script setup lang="ts">
-import Invitations from '@/components/contacts/Invitations.vue';
+import Invitations from '@/components/connections/Invitations.vue';
 </script>
