+++ services/tenant-ui/frontend/src/views/tenant/Profile.vue
@@ -1,33 +1,49 @@
 <template>
-  <h3 class="mt-0">{{ $t('tenant.profile.tenantProfile') }}</h3>
-
-  <div v-if="loading" class="flex justify-content-center">
-    <ProgressSpinner />
-  </div>
-  <div v-else-if="tenant">
-    <div class="grid">
-      <div class="col-fixed w-30rem">
-        <ProfileForm />
-        <Issuance />
-      </div>
-      <div class="col text-right">
-        <img v-if="isIssuer" src="/img/badges/issuer.png" />
-      </div>
+  <MainCardContent
+    :title="$t('tenant.profile.tenantProfile')"
+    :refresh-callback="reloadProfileDetails"
+  >
+    <div v-if="loading" class="flex justify-content-center">
+      <ProgressSpinner />
     </div>
-    <hr class="mb-4" />
-    <ProfileFooter />
-  </div>
+    <div v-else-if="tenant">
+      <div class="grid">
+        <div class="col-fixed w-30rem">
+          <ProfileForm />
+          <Issuance />
+        </div>
+        <div class="col text-right">
+          <img v-if="isIssuer" src="/img/badges/issuer.png" />
+        </div>
+      </div>
+      <hr class="mb-4" />
+      <ProfileFooter />
+    </div>
+  </MainCardContent>
 </template>
 
 <script setup lang="ts">
 import Issuance from '@/components/profile/issuance/Issuance.vue';
+import MainCardContent from '@/components/layout/mainCard/MainCardContent.vue';
 import ProgressSpinner from 'primevue/progressspinner';
 import ProfileForm from '@/components/profile/ProfileForm.vue';
 import ProfileFooter from '@/components/profile/ProfileFooter.vue';
+import { useToast } from 'vue-toastification';
 // State
 import { storeToRefs } from 'pinia';
 import { useTenantStore } from '@/store';
+const tenantStore = useTenantStore();
 const { tenant, isIssuer, loading } = storeToRefs(useTenantStore());
+
+const toast = useToast();
+
+const reloadProfileDetails = async () => {
+  try {
+    await tenantStore.getIssuanceStatus();
+  } catch (error) {
+    toast.error(`Failure getting Issuer info: ${error}`);
+  }
+};
 </script>
 
 <style scoped lang="scss">
